feat: allow the scripts to be overriden
The interpretor can be overriden in addition to the file path
Additional environment variables can be added to allow the scripts to use profiles and aws tokens that are only applicable in the scripts

+++ modules/tf-aws-open-next-s3-assets/scripts/delete-folder.sh
@@ -1,5 +1,32 @@
 #!/bin/bash
 
+## Setup
+
 set -e
+
+## Validation
+
+exitcode=0
+
+if [ ! -x "$(command -v aws)" ]; then
+   exitcode=1
+   echo "Error: AWS CLI not found. See https://docs.aws.amazon.com/cli/latest/userguide/getting-started-install.html to install the CLI"
+fi
+
+if [ -z $BUCKET_NAME ]; then
+   exitcode=1
+   echo "Error: bucket name was not supplied"
+fi
+
+if [ -z $FOLDER ]; then
+   exitcode=1
+   echo "Error: folder was not supplied"
+fi
+
+if [ $exitcode -ne 0 ]; then
+   exit $exitcode
+fi
+
+## Script
 
 aws s3 rm s3://$BUCKET_NAME/$FOLDER/*
